Fix: Se corrigio problemas de importaciones dentro de la pagina de resumen para evitar conflictos con el test de github
